+++ HISTORY.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 ## 4.0.0 - unreleased
 
+* BREAKING changes
+  * Removed support for PHP `<8.0` ([#91] via [#250])
+  * Removed support for Composer `<2.3` ([#153] via [#250])
+  * CLI
+    * Changed option `output-fil` to default to `-` now, which causes to print to STDOUT (via [#250])
+    * Removed option `exclude-dev` in favour of new option `omit` (via [#250])
+    * Removed option `exclude-plugins` in favour of new option `omit` (via [#250])
+  * Dependencies
+    * Requires `cyclonedx/cyclonedx-library:^2.0`, was `:^1.4.2` ([#128] via [#250]) 
+* Added
+  * Support for CycloneDX Spec v1.4 (via [#250])
+  * CLI:
+    * New option `omit` (via [#250])
+    * New switch `validate` to override `no-validate` (via [#250])
+
+[#91]:  https://github.com/CycloneDX/cyclonedx-php-composer/issues/91
+[#128]: https://github.com/CycloneDX/cyclonedx-php-composer/issues/128
+[#153]: https://github.com/CycloneDX/cyclonedx-php-composer/issues/153
+[#250]: https://github.com/CycloneDX/cyclonedx-php-composer/pull/250
+
 ## 3.10.2 - 2022-09-15
 
 Maintenance Release.